fix: correct Linux file descriptor type and ensure compatibility with poll()
- Updated `PollFd.fd` type to `int` to align with Linux `pollfd` struct specification.
- Adjusted file descriptor retrieval and conversion in `HasPendingEvents()` for proper usage with `poll()`.

+++ Yubico.Core/src/Yubico/PlatformInterop/Linux/Libc/Libc.Interop.cs
@@ -117,7 +117,7 @@ namespace Yubico.PlatformInterop
         [StructLayout(LayoutKind.Sequential)]
         internal struct PollFd
         {
-            public IntPtr fd;
+            public int fd;
             public short events;
             public short revents;
         }
